+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - Unreleased
+
+### Changed
+
+- Increase Dbatools module requirement v2.8.2
+- Update Release.yml to migrate from 2022 to latest
+- Update github action dependencies to latest version
+- Increase AsBuiltReport.Core modules to v1.6.4
+- Reorganize module structure - moved module files to AsBuiltReport.Microsoft.Windows/ subdirectory
+
 ## [0.5.7] - 2025-07-30
 
 ### Changed